Populate incoming connections and much more
This change is a big one. Its primary function is to populate incoming
connections for both socket-to-socket (already on main) and
prop-to-prop relationships. Outgoing connections are dropped as they
will be processed in the webworker.

Introducing WeakReferences:

This change introduces the concept of "WeakReferences" in the MV world.
Essentially, weak references allow for a reference to another MV
without adding the referenced MV's checksum into the equation. This
means that re-computation of an existing MV will not recompute the
source MV.

This is most useful for scenarios where you want a field from another MV
that can change at any point in time, but is not entirely relevant to
the MV you are working on (e.g. show me the name of this Component in
the connection).

No More MVs in "dal":

Ever change one line of an MV file and you are waiting minutes upon
minutes for the dal? Why does the dal even have access to MV types
anyway? <to be continued>

Dropped Dependencies:

- Frigg no longer <to be continued>

Signed-off-by: Nick Gerace <nick@systeminit.com>