Populate incoming connections and much more
This change is a big one. Its primary function is to populate incoming
connections for both socket-to-socket (already on main) and
prop-to-prop relationships. Outgoing connections are dropped as they
will be processed in the webworker.

Populating Incoming Connections:

While socket-to-socket connections are already on main, this change
adds prop-to-prop connections. Since outgoing connections will be
inferred in the webworker due to merkle tree hash calculations and the
post-rebase design principles of edda they have been temporarily
dropped.

Introducing WeakReferences:

This change introduces the concept of "WeakReferences" in the MV world.
Essentially, weak references allow for a reference to another MV
without adding the referenced MV's checksum into the equation. This
means that re-computation of an existing MV will not recompute the
source MV.

This is most useful for scenarios where you want a field from another MV
that can change at any point in time, but is not entirely relevant to
the MV you are working on (e.g. show me the name of this Component in
the connection).

No More MVs in "dal":

Ever change one line of an MV file and you are waiting minutes upon
minutes for the dal? Why does the dal even have access to MV types
anyway? If you've asked these questions, this matters to you. Now, the
MV types and MV macros are in their own crates, exclusive to the
edda and dal integration test toolchains.

Dropped Dependencies:

During the refactor, unused dependencies in frigg and si-db triggered
rebuilds when working with MVs for the dal. Now, the dal (and those
crates) should be unaffected by MV changes.

Dependent Feature Flags:

With this change, "app/web" supports dependent feature flags formally.
If you enable NEW_HOTNESS, you'll get all the features you need to make
it work.

Misc:

There are several misc changes, one of which includes optimizing the
"VariantType" MV field to be pascal case rather than having individual
serialization overrides.

Signed-off-by: Nick Gerace <nick@systeminit.com>