fix: don't name NATS Ordered Consumers
This change removes explicit names when creating [Ordered Consumers].
Ordered consumers are ephemeral and the NATS server may re-create these
consumers often server-side. If a name is provided in the consumer
config, it's likely that the NATS server attempts to re-use the
user-provided name which results in a consumer create error.

The fix here is to not explicitly name the consumers and allow the
client and server to coordinate when re-creating consumers is
appropriate. Note that in the Rust client, that field type for
`OrderedConfig.name` is `Option<String>` with the default value being
`None`. As a result we can omit setting this field and rely on the
defaults to get us what we need.

[Ordered Consumers]: https://docs.nats.io/using-nats/developer/develop_jetstream/consumers#ordered-consumers

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>