Add all weak references for incoming connections
Before this change, the only weak reference for incoming connections
was on the "from" side of a "socket-to-socket" connection. Now, the
components for both sides of both connection kinds are weak references.

This change also fixes the "from" side of the "socket-to-socket"
connection to ensure that the component is correct.

All weak references are now created using a "From" impl too for ease of
use.

Signed-off-by: Nick Gerace <nick@systeminit.com>