fix(rebaser): copy & delete errored Rebaser request messages on errror
This change changes the error behavior of the `ChangeSetProcessorTask`
in the Rebaser--that is the task that consumes and processes Rebaser
requests for a change set. Prior to this change, when an early
`Result::Err` is returned from a message processing handler function,
this message was not deleted from the stream and therefore could be
re-processed multiple times.

With this change, when a handler returns with an error a new
`MoveMessageOnFailure` callback in the `PostProcessLayer` will make a
copy of the message into a `rebaser.errored.requests.$wk_id.$cs_id`
subject and then will delete the original message from the requests
stream.

There is little value in re-processing a request, especially if that
request successfully ran through the `perform_rebase()` function. As
before, an `error!` message will be displayed so that we can monitor any
aberrant behavior.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>