fix(rebaser): don't error out task when Pinga returns error from DVU
This change updates the behavior of the `SerialDvuTask` when an error is
returned from Pinga when executing a dependent values update (i.e. DVU)
job. Prior to this change any errors in setting up the job request to
Pinga and job results reported back would cause the `SerialDvuTask` to
return a `Result::Err`, causing the Rebaser change set task to initiate
a shutdown and reschedule of the task work.

This change treats the `SerialDvuTask` as a trigger point to run DVU
jobs and a coodinator, but it does not propagate the errors back from
Pinga and into the Rebaser's change set task. Now the `run_dvu()`
internal function does not return an error but instead will log an
`error!` when something may have gone wrong when triggering a DVU job.
The return of the `ctx.blocking_commit_no_rebase()` method is checked to
see if an execution error has occurred, and in that case it logs a
`warn!` message and continues on.

The principle here is that by prematurely ending the `SerialDvuTask`
when a Pinga error is encountered, the Rebaser is simply propagating
errors across services for circumstances out of its direct control.
These changes attempt to make the Rebaser less responsible for the
problems of others ;)

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>