Add mjolnir smoke test for tools tests
This change adds a mjolnir smoke test for tools crons and general
testing. This will be in the prod tests too.

Why use diagram calls for the test if it is likely going away in favor
of the new map view? The test that was copied (the "2" API test) uses
the same routes.

The README for the sdf API tests has been adjusted too for present day
usage.

Signed-off-by: Nick Gerace <nick@systeminit.com>