Only mark MV kinds as "building" if that kind has been dispatched or queued
In the MV kind dependency graph, we mark a kind as "running, but not
finished yet" by introducing a cycle on itself to prevent trying to
dispatch kinds that depend on it and to prevent trying to dispatch it
again while we're waiting for its tasks to run. We should only mark a
kind if we have actually dispatched, or queued its tasks.

Previously, we could have marked a kind as running, even though we had
not dispatched, or queued its tasks if its dependencies had finished,
but there were still queued tasks to dispatch, which would have
prevented that kind and all downstream kinds from ever being dispatched.