Remove info log of all build tasks of a particular MV kind finishing in edda
This log message was happening multiple times within a single build
batch, because it was only accounting for currently running tasks for
that MV kind, not for any tasks that have been queued to dispatch.

Since edda's logging is pretty noisy at the moment, this is being
removed instead of fixed for that to try to get a better signal to noise
ratio.