Sort adding reference dependencies by MV kind in edda
The order the dependencies for each MV kind is added to the dependency
graph edda uses is not significant on its own. Because of this, we can
sort the order we do this in to make it easier to scan and find if a
particular MV kind is being handled.