Always return the FrontendObject when building MVs
Previously we were skipping returning both the patch from the previous
version, and the new version of the FrontendObject if the new & old
versions had the same checksum to cut down on the number of "empty"
patches we would send to listening clients. However, this resulted in an
unstable MvIndex if any of the MVs had been previously built and we were
doing a full rebuild for any reason. The full rebuild only includes the
FrontendObjects that are returned from building the MVs without
includeing anything from the old/current MvIndex, and when individual
MVs are built, they always check the old/current MvIndex to see if there
is a version that they should generate a patch against, returning
neither a patch, nor the FrontendObject if the generated FrontendObjects
were the same.