Add compiling spike of compressing edda messages
This change adds a compiling spike of compressing edda messages. It is
not functional nor should be merged, but it is a compiling base for
processing "compressed" NATS JetStream messages and distributing work
to handlers accordingly.

It contains a "spike within a spike" to see if we can make naxum
generic to non-NATS messages. Unfortunately, the juice is likely not
worth the squeeze and it may be easier to 1) make a fake NATS message or
modify the last processed one or 2) do not use a naxum server for the
change set processor task and spawn MV builds via a task tracker.

The good news is that the core logic in "compressor.rs" is mostly close.
The idea is that if you see a continguous stream of messages (i.e. the
new "from" snapshot was the last "to" snapshot) that are all "update"
messages, we should try to compress them. The compressed message
contains de-duped changes (WHICH ARE OKAY BECAUSE THEY ARE NOT UPDATES)
that can be used for MV builds accordingly.

There's excess code in this commit that we can likely trash. For
instance, the API types additions in "edda-core" can likely get thrown
out if we decide to not go through NATS because the request type is
entirely in-memory.

My 0.02 cents from this spike: we have to decide what is easier and
better for the system from two options. The options are 1) get naxum to
work with non-NATS messages (or fake it) or 2) do not use naxum for the
change set processor task and use traditional streams. Yes, this is
written similarly to my message earlier about the "spike within a
spike", but you can see that I was contending with this problem. The
hacker in me leans towards cutting out naxum for the change set
processor task and just double acking every message that was
successfully compressed. I fear that the lift to make naxum generic is
not worth it when we already have control over the stream.

At any rate, this change should at least help move the needle forward
for this effort. Follow the "FIXME" comments in this change for more
details.

Signed-off-by: Nick Gerace <nick@systeminit.com>