chore(si-layer-cache): skip disk caching in DAL integration tests
This change disables the disk caching layer in Foyer when running the
DAL (and SDF) integration tests. The si-layer-cache test suite still
enables full disk caching to ensure it continues to operate as before.

Without this configuration update the amount of data written to the
`/tmp` filesystem is enourmous, grows with each new `Cache` entry and
consumes a high number of open file handles in the process. It makes
having the `/tmp` filesystem on `tmpfs` almost unusable (due to disk
consumption needs of the test suites) and cleanup on disk-backed `/tmp`
filesystems takes an incredibly long time due to the large number of
inode records.

Addditonally, the default location for the disk cache was hardcoded to
by under `/tmp` which means that no usage of `$TMPDIR` could override
this decision. This change also changes the default root directory to
use Rust's `std::env::temp_dir()` location which honors the `TMPDIR`
environment variable. This means that setting `export TMPDIR=$(pwd)/tmp`
would cause all temp files for layer cache to be written under the
`./tmp` directory.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>