Treat index pointer deserialization errors the same as an index not existing
In the case where we fail to deserialize the index pointer value the
quickest path to recovery is to pretend as though the index pointer was
not there at all, and the most likely reason for failing to deserialize
the pointer value is that we changed something about what data is in the
value. If we have changed what's in the value, we likely need to
re-generate at least the pointer, if not all of the MVs & the MvIndex as
well, and the easiest way to make sure that happens is to have the
system act as though the pointer was never set.