fix(sdf,web):Decrease risk of race condition, handle removed nodes, and reactive func runs
This doesn't solve it completely, but by returning the current index's snapshot address, we *slightly* reduce the risk that the front end has a newer snapshot than the patches. 

When nodes are removed, we try to look up the previous object in Frigg to get the last checksum. This was causing patches with deletes to not land in some cases. 

Func Run list is now reactive to Func Runs in the grid view!