Create definition checksums for types that derive FrontendChecksum; include them in overall MV definition checksum
Previously, we were not detecting if the structure of a type used by a
materialized view changed, if the type being used was not also deriving
`MV`.

All types included in a materialized view (at any depth) must also
implement `FrontendChecksum`, so the derive macro for `FrontendChecksum`
now also provides the same kind of "definition checksum" registration as
we do when deriving `MV`.

Whenever possible, the implementation of `FrontendChecksum` for a type
should be derived, instead of manually implemented, so that it is also
automatically included in the overall definition checksum of all types
used by materialized views.

This also sets up a static `LazyLock` around
`materialized_view_definitions_checksum()` to memoize the result so we
only have to calculate it once for the lifetime of the process.