feat(luminork): Adding a search endpoint for luminork
This endpoint will take a search payload - the first parameter is schemaName.

We will first find all of the components in the workspace. If we don’t pass a schemaName then we will return the full list of the components, if we pass a schemaName then we will filter the components that match only that schema

Next up will be to return components that can be upgraded!