Ignore errors retrieving index pointer values when attempting to re-use existing MvIndex for a new change set
If we're unable to get the index pointer value for an existing index
pointer then we should ignore it and move on to trying the next one. The
read could fail if we have changed the format of the pointer value, and
this should not cause the entire operation to fail. We should move on,
and act as though the index pointer didn't exist in the first place to
give us a chance of finding an index pointer value that is appropriate
for us to reuse.