Align qualification func run result to its state
In the new UI, the qualification func run state now is reflected by both
a factor of its underlying func run state and the result of said
function. Part of what made this change tricky is the distinction
between "unprocessed result" and "result". The "unprocessed result" is
the one we need for this to work. Tangentially, this should probably
be the result used in the func run details panel too, or at least it
should be accessible to the user somehow.

In addition to these changes, the func kind is now the correct enum
for the func run object.

Signed-off-by: Nick Gerace <nick@systeminit.com>