fix(edda, web): Use Workspace Id for MV list items, rather than change set
This change allows us to actually reuse MV Indexes across change sets. Without this change,when we copy the Index Pointer for a new change set, we never build the list items, and their Ids are pinned to the change set id from which we copied, making them unusable. By keying off of the workspace Id, we can share these objects across change sets, reducing the time to build, time to create change sets, and we get more shared state.