chore(tilt): add "workspaces" link in `web` service
This change adds a "workspaces" link in the Tilt UI under the `web`
service. Prior to this change, navigating to the "web" link for a newly
brought up dev stack causes authentication redirects often sending the
developer to their default workspace rather than the localhost stack. By
having the link to the list of workspaces there's far less navigating
around and causing other workspaces to load in production environments,
etc.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>