chore(edda): add `CompressingStream` telemetry
This change adds a `Span` for each call to `CompressingStream.next()`,
which pulls a `CompressedRequest` into a Naxum app.

Note that the time a span is open involves sitting and waiting for
NATS messages on its subscription and possibly waiting for the Naxum app
to poll the future to pull in the next unit of work (remember that Edda
only performs one request at a time per-change set). There's also a
quiescent period shutdown for this change set processing task, so some
of these spans last for ~60 seconds and then close without yielding
work (the task is being shutdown and so this stream is also being
dropped and closes any spans in-flight).

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>