fix(dal): DVU now enqueues relevant update actions if not on head
This PR shifts some logic from an SDF route handler to the DVU job.  Previously, updating a prop for a component, we would build the DVU graph, and walk it to see if any downstream components might be impacted - all before returning to the caller.  This causes the response time to increase as it becomes more expensive to build the DVU graph. 

Now, this happens in DVU, where we already have all of the data we need. Better yet, this is more accurate than what we had before, as before we would enqueue the update function if a value is impacted. However, at that time, we don't yet know if the value will *change* as a result of the update made, we only know that it *might*. This lets us more precisely detect if a root/domain prop has actually changed before we look to enqueue an update function.

Special casing this in DVU isn't ideal, but will solve the problem for now as we look for a better solution to handle these effects of change.