Add apply change set to new UI
This change adds the ability to apply change sets in the new UI. This
does not include approvals, but leaves the door open to them. The apply
button follows the same design style as "Add a component".

The modal supports keyboard input, including "ESC" for exiting the modal
and "CMD + ENTER" for applying the change set. The modal also follows
a similar style to the rest of the new UI, trading icons for keyboard
shortcut suggestions.

The actions list within the modal is now driven by props and contains
pill counters for each kind of action seen in the list. The pill counter
building block component has been extended to allow slotted-in icons.

Co-authored-by: Wendy Bujalski <wendy@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>