Use SAH Pool VFS for sqlite DB in webworker
The SyncAccessHandle VFS is the highest performance option of the
OPFS-backed configurations (currently) described in the SQLite WASM
documentation.

https://sqlite.org/wasm/doc/trunk/persistence.md#vfs-opfs-sahpool

In local testing switching to the SAH VFS from the standard OPFS
configuration significantly reduces the time from hitting refresh in the
browser to having a fully rendered component list in the new UI when all
of the necessary data is already loaded into SQLite, as well as reducing
the time needed to do the initial import of data into SQLite. In the
warm-sqlite-cache case it was reduced from ~60 seconds to ~20 seconds in
local testing.