fix(edda): impl a liveness-aware quiescent period for compressing stream
This change alters the "quiescent_period" timeout on
`ChangeSetProcessorTask`s by basing the quiet period on the last time
the `CompressingStream`'s state machine was updated. This way, if there
is still work to be done inside the compressing stream, a quiet period
will *not* be detected until we are waiting for another "first message"
(or if it stalls on another intermediate state).

Additionally, the current value of the stream's `State` is recorded as
a `task.state` attribute on the `compressing_stream.next` span. This
would allow us to alert on the compressing stream closing in unexpected
states.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>
Co-authored-by: Jacob Helwig <jacob@systeminit.com>