Add action toggles to context menu
This change adds action toggles to context menu. The context menu will
only be populated with actions if one component is selected.

Known bug: selecting a toggle causes the menu to close. According to
the right click menu from the modeling diagram and the inner components,
this should not happen. However, we can fix this as more things come
into the context menu.

Signed-off-by: Nick Gerace <nick@systeminit.com>