fix(dal): Ensure we only set a qualification check status if there are active validations
Previously, the call to `Qualification::new_for_validations` would ALWAYS retun a top level prop validation even if there were no validations for the component. This now changes the way this works so that we only return a top level prop validation if there are validations - otherwise we skip it!

This now means that Generic frames don’t get a value of 1 beside their qualifications - even though there are none. The same for Credentials!