Add ability to delete and restore components
This change adds the ability to delete and restore components with
ghosting, context menu awareness, and a new audit log kind. It is a
stopping point in making this feature feel great rather than an endpoint
though. Therefore, I've listed some "known issues" and "planned
improvements" at the bottom of this description.

The primary result of this change is a new deletion modal that handles
both deleting components and removing them from views (which includes
new logic for "color" derivation from an category name). The new logic
relies on the "toDelete" field of a component, which is now returned
via the Component MVs.

Component Context Menu Changes:

Along the way, the component context menu now uses full "Component"
objects rather than just the IDs. This is so that we can populate the
modals with information as need be. There's more relevant changes
too...

- Handle multi-select component edge cases (e.g. both ghosted and
  regular components have been selected)
  - This is not in the UI today, but the Vue component could become
    very frustrating to use quickly without some safeguards upfront to
    keep us moving
- Co-locate all "single component" logic
- Make "DeleteModal" the first user of full "Component" objects
- Standardize on a components "ref" rather than going through both
  "props" and "ref" for components
- Provide "BifrostComponentInList" with a "schemaVariantId" for
  information population

Known Issues:

- No reactivity on the grid when deleting and restoring components
  (hitting "re-do" cold start fixes it)
  - Investigated with @jobelenus and it looks like the "node_hash" is
    being calculated correctly... something is weird is happening here
- Clicking the restore context menu option does not close the context
  menu

Planned Improvements:

- Showing the full component details, but putting everything in a
  read-only state
- Showing the full component details and showing a banner on what a
  ghosted component is as well as a button to restore it
- Showing the full word "Restore" rather than "R" for the context menu
  restore option
- Centering the hourglass icon to the pill counters in a ghosted grid
  tile title

Signed-off-by: Nick Gerace <nick@systeminit.com>