Remove cross-MV dependency ordering when building MVs
The intent of having strong references determine the build ordering of
MVs was that MVs were going to have access to other MVs that had been
built earlier within the same batch. This would have allowed them to
know the checksum of the other MVs without having to build it for
itself, which would duplicate the amount of work performed by edda.

Instead, we have been moving towards using weak references (only the
kind & ID, with no checksum), which do not require either having access
to MVs that may have been built within the same batch, or to re-built
the MV every time it is going to be referenced from another MV.

Rather than limiting which MVs can be built at the same time in
preparation for a direction we are moving away from, the only limit to
the parallelism is now the `PARALLEL_BUILD_LIMIT`, and all MVs are
assumed to be able to build without the output from any other MV being
available (which is the current state of all MVs).