Change last remaining strong Reference in MVs to be a WeakReference
The `IncomingConnectionsList` was the only remaining `Reference`, and
creating the `Reference` instead of a `WeakReference` meant having to
build the `Component` MV for every component an additional time to be
able to have the checksum necessary for the `Reference`.

Additionally, the front end already has enough information to be able to
deterministically resolve any `WeakReference` to a specific checksum
because of the `MvIndex`.