feat(web): use a shared webworker to enable sqlite3 shared access pool
Uses weblocks to ensure that only one tab's worker task has access to
the sqlite3 database at a time. Proxies requests through a SharedWorker
to the worker that holds the lock. As soon as a tab closes, all tabs
will race for the lock, and the one that acquires it first will send
a proxy of itself to the SharedWorker.

Requires all DBInterface methods that hit the database to be async (so
they can wait for a tab remote to be ready)