fix(web): Skip initializing bifrost to prevent multiple cold starts when switching between old and new experience
We had previously exposed some data to the existing user experience using the new architectural pattern, behind feature flags. The flags have since been disabled with no intention of reenabling them, and when navigating from the diagram to the grid view, this is causing bifrost to initialize twice - one is enough!

Note: I am not going as far as ripping out all of the unexercised code behind those flags - not worth it right now.