Add ability to edit and delete views
This change adds the ability to edit and delete views with protections
against non-empty views. The existing "AddViewModal" was used as a
frame of reference despite the new approval modal being closer to the
desired designs.

Some complications in here included adding the "settings-edit" button,
which behaves differently than "endLinkTo" in "DropdownMenuItem" as well
as handling "canDeleteView" and query timing. Leveraging the
"beforeOptions" and "afterOptions" for the button as well as allowing
options to expand past the max width were part of getting here. This one
was painful to get close, but we got there.

Signed-off-by: Nick Gerace <nick@systeminit.com>