Try to reduce nats load by only inserting FrontEndObjects if they're not already in the KV store
This doesn't reduce the amount of information sent over the wrire to
nats on an insert, but nats should be able to short-circuit and avoid
doing any write in the case where the object has already been generated
for any reason.

With the interface provided by the nats KV store, our options for only
inserting if the key doesn't already exist looked to be:

* Send the (potentially) new thing over the wire and watching for a
  failure (`create` behaves this way).
* Try to get first (and receive the full value).
* Iterate over all of the keys to see if the key is already there.

Trying the insert using `create` and ignoring the "already exists" error
kind seems to be the least-bad of the options.