Initial sketch of streaming patches to the front end from edda
The idea with streaming patches to the front end is that it will be able
to start using the new MVs as soon as they are generated, before _all_
of the data is ready, and there will be a message with the MvIndex patch
(message not created/sent yet) that will let the front end know that the
stream of patches has finished, and which objects it should have by the
end (fetching any it missed the patch for).

Right now, these streaming patches are sent in addition to the existing
patch batches, but once we start using the streamed patches in the front
end, we'll want to stop creating & sending the batched patches as it
will be purely redundant from the streamed patches.

We will also want to figure out if the existing `IndexUpdate` message
should be augmented to include both the from-checksum & patch for the
MvIndex, or if we want an entirely new message to handle this.