Protect the default view from deletion
This change protects the default view from deletion at the graph level
as well as provides tooltip text for default views in the edit modal in
the new UI.

In addition to that change, the local development docs have been edited
to mention the SSL requirement when not using localhost. This was
already recommended, but is now required with the recent webworker
changes and its lock acquisition code.

Signed-off-by: Nick Gerace <nick@systeminit.com>