fix(edda): ComponentList should have a weak reference to ComponentInList
This corrects the weak reference pointer to be the ComponentInList. The impact of the incorrect pointer meant that when tanstack queried for the ComponentList MV, it was hydrated with the Component MVs as opposed to the ComponentInList MVs. This lead to the upgrade option in the context menu being added to all components, regardless of upgrade status.