wip: rbe
This tries (and fails) to make remote builds work with this solution.

The core problem here is that rust generally relies on clang and friends
to be available on the PATH, so we need to include it in the toolchain.
However, the default clang tool from mise won't build without cmake and
ninja. Including those as packages doesn't make it build. I could use a
container image that has them, but that feels like it defeats the point.

There is an alternative clang you can use, but it doesn't include a symlink
from clang++ to clang and if you make one manually, mise just gets mad
because clang++ is not one of its shims. We are generally just brushing
up against the nature of hermetic builds. It's hard to buid the entire
environment on the fly. I tried to replicate what this PR does by
pulling bins out of the nix flake, and that actually almost worked, but
nix dynamically links everything, so you are either shipping the entire
flake or rebuilding statically.

I think this pattern is close, but I am not entirely sold on mise being
a viable tool here. The lift of it replacing the nix flake seems pretty
huge and I doubt we can continue on without both if we decide to use
mise.