chore(dev): move Prometheus config & simplify config
This change moves the location of the development Prometheus config to
the normalized directories like the other telemetry services and removes
the exposed port to the host. As Grafana is the entrypoint to consume
Prometheus data and due to the overly ubundant occurances of
Prometheus-speaking services (which all tend to bind to port 9090), it
seemed simpler to not expose this port on the host.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>