fix(edda): Ensure new management edges propagate to the front end
The IncomingConnections MV currently includes both management connections and prop to prop connections.  The management edges were backwards in this MV, which meant that when a new management edge was created (Component A -> Manages -> Component B), the component that gets triggered to rebuild is the one newly under management (Component B), only when building this MV, we're asking about the managed components (Component B -> Manages -> GO FIND THEM), not the managing components (GO FIND THEM BECAUSE THERE IS A NEW ONE NOW -> Manages -> Component B).